♻️ useData hooks로 리팩터링
